refactor: move create mock function to mocks pkg to avoid util pkg
