+++ internal/store/mocks/mock.go
@@ -1,7 +1,8 @@
-package util
+package mocks
 
 import (
 	"database/sql"
+	"fmt"
 
 	"github.com/DATA-DOG/go-sqlmock"
 )
@@ -9,7 +10,8 @@ import (
 func CreateMock() (*sql.DB, sqlmock.Sqlmock, error) {
 	db, mock, err := sqlmock.New(sqlmock.QueryMatcherOption(sqlmock.QueryMatcherEqual))
 	if err != nil {
-		return nil, nil, err
+		return nil, nil, fmt.Errorf("create mock: %w", err)
 	}
+
 	return db, mock, nil
 }
